refactor(images): extract provider dispatch service (#1855)
